Refactor hid module
Move hid module into separate files in preparation of mouse support.
